pyproject.toml
[tool.poetry]
name = "fitness-assistant"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]  # Замени на свои данные
readme = "README.md"




[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.9"
sqlalchemy = "^2.0.38"
pydantic = "2.10.6"
pydantic-settings = "2.8.0"
python-dotenv = "1.0.1"
uvicorn = "0.34.0"
uvloop = "0.21.0"
langchain = "^0.3.19"
langchain-openai = "^0.3.7"
alembic = "^1.15.1"
asyncpg = "^0.30.0"
httpx = "^0.28.1"
psycopg = { version = "^3.1.18", extras = ["binary"] }  # ✅ Обновлено
aiogram = "^3.4.1"
langchain-google-genai = "^2.0.11"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
